Fix collision issue with detachable joints (#31)
Fix collision issue with detachable joints

The issue occurs if a model is spawned and is attached to a parent model
without a call to `step()` in between. This should be fixed in DART, but
in the meantime, we can just call `incrementVersion` after `moveTo` is
called on a DART BodyNode.