Fix collision detection issue when using joints across nested models (#268)
When a joint is created between BodyNodes in different skeletons, the child BodyNode is moved to the skeleton of the parent BodyNode. When this happens, the BodyNode version of the child needs to be incremented. This is actually fixed by dartsim/dart#1445, but we don't have that merged into our fork. So in the meantime, we call incrementVersion after moveTo is called similar to #31.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

Co-authored-by: Louise Poubel <louise@openrobotics.org>