dartsim: add data to LinkInfo needed for loops
In gazebo classic, kinematic loops are closed by
identifying a link with multiple parents and
creating a mirror body node that is welded to
the original. The inertia will be split evenly
between the body nodes. This change adds data
to store the welded mirror nodes, the weld
constraints, and the original inertial data.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>