Add bullet-featherstone plugin (#373)
This PR adds a gz-physics-bullet-featherstone-plugin which wraps the Featherstone API of the bullet physics engine. This is created as a separate plugin from gz-physics-bullet-plugin because the API of Bullet's Featherstone dynamics is different enough that there isn't really feature parity or easy interchangeability between them. Users will need to choose whether they are willing to settle for fewer features in exchange for the better performance of Bullet's Featherstone implementation.


Signed-off-by: Louise Poubel <louise@openrobotics.org>
Signed-off-by: Michael X. Grey <grey@openrobotics.org>
Signed-off-by: ahcorde <ahcorde@gmail.com>
Signed-off-by: Michael Carroll <michael@openrobotics.org>
Co-authored-by: Louise Poubel <louise@openrobotics.org>
Co-authored-by: Alejandro Hernández Cordero <ahcorde@gmail.com>
Co-authored-by: Steve Peters <scpeters@openrobotics.org>
Co-authored-by: Michael Carroll <michael@openrobotics.org>