Fix memory corruption due to faulty refcount tracking (#480)
The memory corruption seems to occur only on DART 6.9.2 (version that comes with Ubuntu 20.04), but not 6.12 (Ubuntu 22.04), so I'm led to believe it's a bug in dart that has since been fixed. But the patch in this PR seems to avoid the problem by not storing the freejoint created when creating links. My suspicion is that this freejoint is not properly taken into account when removing skeletons.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>