Add test with fast/slow pendulum mimics
Add a test world with 3 pairs of fast and slow pendulums.
Onee pair is uncoupled, and the other pairs are mimicked
to each other with multiplier = 1, one with
fast following slow, and one with slow following fast.
The test expects that the mimicked pendulums move at the
same rate, at a speed in between that of the slow and
fast pendulums.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>