bullet-featherstone: Fix bounding box for collisions with pose offset (#647)
Currently if a collision has a pose offset, the bullet-featherstone plugin returns an incorrect axis aligned bounding box for the link containing the collision. This is found to be caused by incorrect frame data pose computed for collisions - It currently ignores the collision pose offset.

This PR makes a few changes to KinematicsFeatures::FrameDataRelativeToWorld in bullet-featherstone plugin:

- fixed getting collision pose relative to world
- fixed getting model pose relative to world
- minor refactoring
The link_features's bounding box test was previously skipped for bullet-featherstone plugin. This PR enables this test for bullet-featherstone by requiring only a minimal set of features needed to run the test.

---------

Signed-off-by: Ian Chen <ichen@openrobotics.org>