Prevent crash when objects move to invalid poses (#706)
DART seems to compute invalid poses for certain objects (e.g. Capsule, Ellipse) with large accelerations. The computed poses end up with `nan` values resulting in a crash. This patch detects this condition and reassigns the poses to the last known poses and resets velocities to zero to prevent the crash.

---------

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>