Fix collision checking issues in bullet-featherstone (#727)
Fixes three issues in bullet-featherstone plugin:

1. After calling SetPosition to set a joint position, incorrect collision transforms are used in the next forward step
2. After calling SetWorldPose to set the pose of a FreeGroup, incorrect collision transforms are used in the next forward step
3. Collisions are ignored between a static object and any link on a fixed-base model that is a child of a fixed joint

Added a GzMultiBody wrapper for btMultiBody to ensure that a flag is set whenever joint position is set or base transform is set indicating that collision transforms need to be updated. Then collision transforms are updated in the next forward step for any body which has the flag set, and the flag is reset.

---------

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>